+++ dbt/adapters/redshift.py
@@ -1,39 +1,16 @@
-import copy
-
-import dbt.flags as flags
-
 from dbt.adapters.postgres import PostgresAdapter
-from dbt.contracts.connection import validate_connection
-from dbt.logger import GLOBAL_LOGGER as logger
+from dbt.logger import GLOBAL_LOGGER as logger  # noqa
 
 
 class RedshiftAdapter(PostgresAdapter):
 
-    date_function = 'getdate()'
+    @classmethod
+    def type(cls):
+        return 'redshift'
 
     @classmethod
-    def acquire_connection(cls, profile):
-        # profile requires some marshalling right now because it includes a
-        # wee bit of global config.
-        # TODO remove this
-        credentials = copy.deepcopy(profile)
-
-        credentials.pop('type', None)
-        credentials.pop('threads', None)
-
-        result = {
-            'type': 'redshift',
-            'state': 'init',
-            'handle': None,
-            'credentials': credentials
-        }
-
-        logger.info('Connecting to redshift.')
-
-        if flags.STRICT_MODE:
-            validate_connection(result)
-
-        return cls.open_connection(result)
+    def date_function(cls):
+        return 'getdate()'
 
     @classmethod
     def dist_qualifier(cls, dist):