+++ test/integration/001_simple_copy_test/test_simple_copy.py
@@ -1,6 +1,7 @@
 from nose.plugins.attrib import attr
 from test.integration.base import DBTIntegrationTest
 
+
 class TestSimpleCopy(DBTIntegrationTest):
 
     def setUp(self):