+++ test/integration/014_hook_tests/test_run_hooks.py
@@ -119,7 +119,7 @@ class TestPrePostRunHooks(DBTIntegrationTest):
         self.assertEqual(ctx['target.name'], 'default2')
         self.assertEqual(ctx['target.port'], 5432)
         self.assertEqual(ctx['target.schema'], self.schema)
-        self.assertEqual(ctx['target.threads'], 1)
+        self.assertEqual(ctx['target.threads'], 4)
         self.assertEqual(ctx['target.type'], 'postgres')
         self.assertEqual(ctx['target.user'], 'root')
         self.assertEqual(ctx['target.pass'], '')