+++ test/unit/test_postgres_adapter.py
@@ -4,7 +4,7 @@ import dbt.flags as flags
 
 from dbt.adapters.postgres import PostgresAdapter
 from dbt.exceptions import ValidationException
-from dbt.logger import GLOBAL_LOGGER as logger
+from dbt.logger import GLOBAL_LOGGER as logger  # noqa
 
 
 class TestPostgresAdapter(unittest.TestCase):
@@ -23,7 +23,8 @@ class TestPostgresAdapter(unittest.TestCase):
 
     def test_acquire_connection_validations(self):
         try:
-            connection = PostgresAdapter.acquire_connection(self.profile)
+            connection = PostgresAdapter.acquire_connection(self.profile,
+                                                            'dummy')
             self.assertEquals(connection.get('type'), 'postgres')
         except ValidationException as e:
             self.fail('got ValidationException: {}'.format(str(e)))
@@ -32,16 +33,7 @@ class TestPostgresAdapter(unittest.TestCase):
                       .format(str(e)))
 
     def test_acquire_connection(self):
-        connection = PostgresAdapter.acquire_connection(self.profile)
+        connection = PostgresAdapter.acquire_connection(self.profile, 'dummy')
 
         self.assertEquals(connection.get('state'), 'open')
         self.assertNotEquals(connection.get('handle'), None)
-
-    def test__get_connection(self):
-        connection = PostgresAdapter.get_connection(self.profile)
-        duplicate = PostgresAdapter.get_connection(self.profile)
-
-        self.assertEquals(connection.get('state'), 'open')
-        self.assertNotEquals(connection.get('handle'), None)
-
-        self.assertEquals(connection, duplicate)